Created docker build for jenkins build. Integration tests will be triggered from master api and admin builds.
