+++ docker/Dockerfile
@@ -1,0 +1,21 @@
+FROM python:3.4-slim
+
+ENV PYTHONUNBUFFERED=1 \
+	DEBIAN_FRONTEND=noninteractive
+
+RUN \
+	echo "Install base packages" \
+	&& apt-get update \
+	&& apt-get install -y --no-install-recommends \
+		make \
+		git \
+	&& echo "Clean up" \
+	&& rm -rf /var/lib/apt/lists/* /tmp/*
+
+RUN \
+	echo "Install global pip packages" \
+	&& pip install \
+		virtualenv
+
+
+WORKDIR /var/project
