+++ integration_tests/schemas/v0/GET_notification_return_email.json
