+++ integration_tests/schemas/v0/POST_notification_return_sms.json
