+++ integration_tests/schemas/v0/email_notification.json
