+++ integration_tests/schemas/v0/sms_notification.json
